backend/pyproject.toml
[project]
name = "agent-ai-backend"
version = "0.1.0"
description = "Realtime Graph Multi-Agent Simulation Report System - Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.12",
    "litellm>=1.50.0",
    "pymupdf>=1.24.0",
    "pyyaml>=6.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "networkx>=3.2.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
